adds ranges::equal_to
